Update tests to try and catch error
Never trust a test you haven't seen fail.
In this case line 88 of BookmarksMemberExtension calls PHP's count() on
what could be `false` (returned from `getTopLevelParent`), causing an
error since PHP 8.
This is not caught in any test case currently, but is affecting real
users in production.

In order to try and understand the use case that would cause this, I'm
writing tests first to ensure I have fixed the issue (not just hidden it
again).