Create a failing test to match PROD error
An issue in production preventing pages from loading after the upgrade
from PHP 7.4 to PHP 8.1 had an interesting setup that was unintuitive to
replicate. Turns out if a page is unpublished but a child that a
bookmark references is still published (some how) this will cause an
error as one of the private helper functions in BookmarksMemberExtension
returns false instead of an array, in turn being passed to count() which
now causes an error instead of a warning.

Now we have a failing test we have a way to prove that our fix is
successful, when we implement one.