Tidy up a bit
Code formatting  quality, dev tool configurations, support translations,
etc.

Minor refactoring in that the member extension cannot be applied to
null, so checking if the member exists is irrelevant. If it is not yet
in the database, then the bookmarks will be an empty list.

There are cleaner ways to set ModelAdmin GridFieldConfig :)

Docblocks should inherit when the declaration is an override, so we can
remove some that are superfluous.