FIX bookmarks with unpubilshed parents breaking site
When an ancestor page is unpublished, but a child page with a bookmark
on it remains published, the member to whom this bookmark belongs
receives a 500 error for any page on the site.

This is due to PHP 8 increasing warning errors to fatal errors with type
expectation mismatches - in this case `false` being passed to `count()`
which expects an array.

Now with a failing test and a more comprehensive test suite we can have
confidence that this is both fixed, and did not introduce another error
by inadvertently altering behaviour.